raspberry_pi_code/requirements.txt
opencv-python
mediapipe
numpy
imutils
RPi.GPIO
